[fe] refactor: context 변수명 변경, 필요없는 import 삭제, 변수명 변경
play 폴더 gameprogression으로 변경
